requirements.txt
faiss-cpu
langchain_community
ipykernel
ollama
langchain-chroma
chromadb
sentence-transformers
streamlit
